// src/store/manage/type.ts
// export interface I

export interface IManageState {
  projectList: any[]
  projectCount: number
  visitorList: any[]
  visitorCount: number
  visitList: any[]
  visitCount: number
  interfaceList: any[]
  interfaceCount: number
  exceptionList: any[]
  exceptionCount: number
  performanceList: any[]
  performanceCount: number
  userList: any[]
  userCount: number
  menuList: any[]
  menuCount: number
  roleList: any[]
  roleCount: number
  project_categoryList: any[]
  project_categoryCount: number
  exception_categoryList: any[]
  exception_categoryCount: number
}